ZoomSolved
the code in these files can take an image, image.jpg and a template, create test cases of the original image zoomed in and out(see sharedRescourses , zoomFactors) with zoomGeneraotor. After that, zoomRecognition takes the template , enlarges it according to the zoomFactor and matches it to the image. All  methods perform well, except method 2, which is always wrong